move firebase config
