+++ src/context/AuthContext.js
@@ -2,7 +2,7 @@
 import React, { createContext, useContext, useEffect, useState } from 'react';
 import { onAuthStateChanged } from 'firebase/auth';
 import PropTypes from 'prop-types';
-import { auth } from '../../firebase';
+import { auth } from '../../firebase/config';
 
 const AuthContext = createContext();
